+++ packages/reactabular/CHANGELOG.md
@@ -3,6 +3,7 @@
 
   * Improvement - edit - Allow `activateEvent` (default `onClick`) to be overridden.
   * Improvement - easy - Trigger `onMoveColumns` only after moving columns has finished.
+  * Improvement - reactabular - New `select` module for handling row selections.
 
 1.1.5 / 2016-08-04
 ==================